modif export dans bash

+++ bin/version/recup_all_version.sh
@@ -5,7 +5,10 @@ FOLDER_ENV_DEF="$FOLDER_DOCKER/file_env"
 
 while read line  
 do   
-   export $line
+   if [ ! -z "$line" ]
+   then
+      export $line
+   fi
 done < $FOLDER_ENV_DEF/.env
 
 ${0%/*}/recup_version_nodejs.sh
@@ -14,9 +17,23 @@ ${0%/*}/recup_version_pm2Logrotate.sh
 
 while read line  
 do   
-   export $line
+   if [ ! -z "$line" ]
+   then
+      export $line
+   fi
 done < "${0%/*}/../../$PROJECT_TMP_MAIN/env_version.txt"
 
+if [ ! -f "$NODEJS_VERSION" ]
+then
 sed -i "s/FROM node:latest/FROM node:$NODEJS_VERSION/" $FOLDER_DOCKER/$DOCKER_FOLDER_PROJECT/Dockerfile
+fi
+
+if [ ! -f "$PM2_VERSION" ]
+then
 sed -i "s/RUN npm install -y --no-install-recommends pm2 -g/RUN npm install -y --no-install-recommends pm2@$PM2_VERSION -g/" $FOLDER_DOCKER/$DOCKER_FOLDER_PROJECT/Dockerfile
+fi
+
+if [ ! -f "$PM2_LOGROT_VERSION" ]
+then
 sed -i "s/RUN pm2 install pm2-logrotate/RUN pm2 install pm2-logrotate@$PM2_LOGROT_VERSION/" $FOLDER_DOCKER/$DOCKER_FOLDER_PROJECT/Dockerfile
+fi
